Deduplicate SendStream
