+++ h3/src/server.rs
@@ -1,10 +1,9 @@
-use std::convert::TryFrom;
-use std::marker::PhantomData;
-
 use bytes::{Bytes, BytesMut};
 use futures::future;
-use http::{response, HeaderMap, Request, Response};
+use http::{response, Request, Response};
+use std::convert::TryFrom;
 
+use crate::connection::RequestStream;
 use crate::frame;
 use crate::{
     connection::{Builder, ConnectionInner},
@@ -34,7 +33,7 @@ where
     ) -> Result<
         Option<(
             Request<()>,
-            RequestStream<FrameStream<C::BidiStream>, Bytes>,
+            RequestStream<FrameStream<C::BidiStream>, Bytes, Connection<C>>,
         )>,
         Error,
     > {
@@ -65,11 +64,7 @@ where
             request_builder
                 .body(())
                 .map_err(|_| Error::Peer("invalid headers"))?,
-            RequestStream {
-                stream,
-                trailers: None,
-                _phantom: PhantomData,
-            },
+            RequestStream::new(stream),
         )))
     }
 }
@@ -85,15 +80,10 @@ where
     }
 }
 
-pub struct RequestStream<S, B> {
-    stream: S,
-    trailers: Option<Bytes>,
-    _phantom: PhantomData<B>,
-}
-
-impl<S> RequestStream<S, Bytes>
+impl<S, C> RequestStream<S, Bytes, Connection<C>>
 where
     S: quic::SendStream<Bytes>,
+    C: quic::Connection<Bytes>,
 {
     /// Send response headers
     pub async fn send_response(&mut self, resp: Response<()>) -> Result<(), Error> {
@@ -109,77 +99,5 @@ where
         frame::write(&mut self.stream, Frame::Headers(block.freeze())).await?;
 
         Ok(())
-    }
-
-    /// Send some data on the response body.
-    pub async fn send_data(&mut self, buf: Bytes) -> Result<(), Error> {
-        frame::write(
-            &mut self.stream,
-            Frame::Data {
-                len: buf.len() as u64,
-            },
-        )
-        .await?;
-        self.stream
-            .send_data(buf)
-            .map_err(|e| Error::Io(e.into()))?;
-        future::poll_fn(|cx| self.stream.poll_ready(cx))
-            .await
-            .map_err(|e| Error::Io(e.into()))?;
-
-        Ok(())
-    }
-
-    /// Send a set of trailers to end the response.
-    pub async fn send_trailers(&mut self, trailers: HeaderMap) -> Result<(), Error> {
-        let mut block = BytesMut::new();
-        qpack::encode_stateless(&mut block, Header::trailer(trailers))?;
-
-        frame::write(&mut self.stream, Frame::Headers(block.freeze())).await?;
-
-        Ok(())
-    }
-
-    // Wait for all data to be sent and close the stream
-    pub async fn finish(&mut self) -> Result<(), Error> {
-        future::poll_fn(|cx| self.stream.poll_ready(cx))
-            .await
-            .map_err(|e| Error::Io(e.into()))?;
-        future::poll_fn(|cx| self.stream.poll_finish(cx))
-            .await
-            .map_err(|e| Error::Io(e.into()))
-    }
-}
-
-impl<S> RequestStream<FrameStream<S>, Bytes>
-where
-    S: quic::RecvStream,
-{
-    /// Receive some of the request body.
-    pub async fn recv_data(&mut self) -> Result<Option<Bytes>, Error> {
-        match future::poll_fn(|cx| self.stream.poll_next(cx)).await? {
-            Some(Frame::Data { .. }) => (),
-            Some(Frame::Headers(encoded)) => {
-                self.trailers = Some(encoded);
-                return Ok(None);
-            }
-            Some(_) => return Err(Error::Peer("Unexpected frame type on request stream")),
-            None => return Ok(None),
-        }
-
-        Ok(future::poll_fn(|cx| self.stream.poll_data(cx)).await?)
-    }
-
-    /// Receive an optional set of trailers for the request.
-    pub async fn recv_trailers(&mut self) -> Result<Option<HeaderMap>, Error> {
-        let mut frame = future::poll_fn(|cx| self.stream.poll_next(cx)).await?;
-
-        let fields = match frame {
-            Some(Frame::Headers(ref mut b)) => qpack::decode_stateless(b)?,
-            Some(_) => return Err(Error::Peer("Unexpected frame types")),
-            None => return Ok(None),
-        };
-
-        Ok(Some(Header::try_from(fields)?.into_fields()))
     }
 }
